Move other_controller up
